+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## 2.10.0
 
+- Deprecate setting minimum_coverage value. Use skip_response_coverage, ignored_unknown_status to configure coverage instead.
 - Update openapi_parameters to make parsing array query parameters more consistent.
   Now parsing empty array query parameter like `ids=&` or `ids&` both result in an empty array value (`[]`) instead of `nil` or `""`.
 - Fix Test::Coverage.result returning < 100 even if plan is fully covered